summarizer/requirements.txt
transformers
datasets
tokenizers
wandb
torchinfo
evaluate
rouge_score
SentencePiece
torchmetrics